SyncServer - refactored delegate location
